pyproject.toml
[tool.poetry]
name = "variantcallingpipeline"
version = "0.1.0"
description = ""
authors = ["Benjamin Dumont <benjamin.dumont294@outlook.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
biopython = "^1.80"
python-dotenv = "^0.21.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
mypy = "^0.991"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
